Update to more recent upstream parent POM and modern Checkstyle rules
Eclipse has included a much more modern version of Checkstyle in recent versions, which requires backwards incompatible changes in Checkstyle configuration and thus a new version of Checkstyle in the Maven context.

The simplest solution to this is to move to a much more modern version of the Shibboleth parent POM and pick up most configuration from there.